Allow users to specify max websocket receive size
::WebSocket::Driver has a limit on the amount of data it is willing to
receive from the remote side. This limit is configurable via the
:max_length option and defaults to 64MB.

One way to hit this limit is to print a large PDF (> 48 MB). Doing so
will cause ::WebSocket::Driver to close the socket with error code 1009
and error message "WebSocket frame length too large", which in turn
causes Ferrum to raise a `Ferrum::DeadBrowserError`.

This commit adds a `:ws_max_receive_size` option which can be used to
change this limit.